requirements.txt
amqp==2.5.0
astroid==2.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.9.1
djangorestframework-jsonapi==2.7.0
djangorestframework-simplejwt==3.3
djoser==1.5.1
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
kombu==4.6.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.9
redis==3.2.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wrapt==1.11.1
